fix : video-layer
fix : video-layer

update Vuex Store currentVideo
divide Video components
