+++ frontend/package.json
@@ -15,7 +15,8 @@
     "three": "^0.145.0",
     "videojs": "^1.0.0",
     "vue": "^2.6.14",
-    "vue-router": "^3.5.3"
+    "vue-router": "^3.5.3",
+    "vuex": "^3.6.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
