+++ frontend/src/App.vue
@@ -5,13 +5,9 @@
 </template>
 
 <script>
-// import Header from "@/components/Header.vue";
-
 export default {
   name: "App",
-  components: {
-    // Header,
-  },
+  components: {},
 };
 </script>
