--- frontend/src/components/Header.vue
@@ -1,12 +1,0 @@
-<template>
-  <div>
-    <router-link to="/analysis">video</router-link>
-    <router-link to="/main">main</router-link>
-  </div>
-</template>
-
-<script>
-export default {};
-</script>
-
-<style scoped></style>
